% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_data_based_molecular_distinction_score.R
\name{bin_data_based_distinction}
\alias{bin_data_based_distinction}
\title{Distinction p values for each gene for each pair of domains}
\usage{
bin_data_based_distinction(bin_data, meta_data)
}
\arguments{
\item{meta_data}{}
}
\description{
Distinction p values for each gene for each pair of domains
}
