% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_bins.R
\name{bining_function}
\alias{bining_function}
\title{Binning Each Cell}
\usage{
bining_function(data, binsize)
}
\arguments{
\item{binsize}{}
}
\value{
bin_data: The binned data where rows are bins and columns are all genes. Each entry (i,j) is the number of transcripts of Gene(j) in Bin(i).

meta_data: The meta data for each bin. This contains uID, centers of each bin (node_center_x, node_center_y), nucleus state (0: inside the nucleus, 1: outside the nucleus, -1: boundary of the nucleus ), and index column
}
\description{
The function that takes transcripts coordinates as inputs and divide each cell into square bins. User needs to specify the binsize.
}
