% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plots_per_domain.R
\name{density_plots_per_domain}
\alias{density_plots_per_domain}
\title{Total number of transcripts in each domain}
\usage{
density_plots_per_domain(transcript_data_with_domains)
}
\arguments{
\item{transcript_data_with_domains}{}
}
\description{
This function returns a boxplot for total number of transcript in each domain. Each point of boxplot is a cell.
}
