% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex_enrichment_scores.R
\name{enrichment_score}
\alias{enrichment_score}
\title{APEX-seq Enrichment Scores}
\usage{
enrichment_score(weights_file, exp_data, meta_data)
}
\arguments{
\item{meta_data}{}
}
\description{
This function calculates enrichment scores using decoupleR package
}
