% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex_enrichment_scores.R
\name{enrichment_score_plotter}
\alias{enrichment_score_plotter}
\title{Enrichment Score Heatmap}
\usage{
enrichment_score_plotter(enrichment_score_df)
}
\arguments{
\item{enrichment_score_df}{}
}
\description{
Enrichment Score Heatmap
}
