% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot_domain_wise.R
\name{mean_expression_cell_wise_pca}
\alias{mean_expression_cell_wise_pca}
\title{PCA of mean expression of each domain from all cells}
\usage{
mean_expression_cell_wise_pca(
  meta_data,
  data,
  color_values = NULL,
  mean_data = NULL
)
}
\arguments{
\item{color_values}{}
}
\description{
PCA of mean expression of each domain from all cells
}
