% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts_data_with_assigned_domains.R
\name{merge_function}
\alias{merge_function}
\title{Assign domain to each trasncript for a cell}
\usage{
merge_function(cell, binned, binsize)
}
\arguments{
\item{binsize}{}
}
\description{
Assign domain to each trasncript for a cell
}
