% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobssd.R
\name{mobssd}
\alias{mobssd}
\title{MoBSSD function}
\usage{
mobssd(
  threshold_data,
  meta_data,
  num_domains,
  neighborhood_list = NULL,
  parameters_initial_list = NULL,
  niter = 200,
  gamma_fixed = FALSE,
  gamma = 0,
  projection_type = NULL
)
}
\arguments{
\item{projection_type}{}
}
\value{
domains

omega

probability_matrix

meta_data_with_domains
}
\description{
This function fits the latent states based model to the threshold data.
}
