% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_for_binning.R
\name{nucleus_marker}
\alias{nucleus_marker}
\title{Title}
\usage{
nucleus_marker(group)
}
\arguments{
\item{group}{A dataframe of coordinates of a cell with distance from nuclear membrane column named as "inNucleus"}
}
\value{
Returns 0 if bins are inside
}
\description{
Title
}
