% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_initialization.R
\name{parameters_init_function}
\alias{parameters_init_function}
\title{Parameters Initialization Function for EM algorithm}
\usage{
parameters_init_function(
  threshold_data,
  meta_data,
  projection_type,
  num_domains
)
}
\arguments{
\item{num_domains}{}
}
\value{
domain_initial

pmat_initial

omega_initial
}
\description{
This function uses UMAP or PCA based clustering approach to provide intilization values for the model parameters.
}
