% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_model.R
\name{potts_model}
\alias{potts_model}
\title{Model for Subcellular Spatial Domain Discovery}
\usage{
potts_model(
  L,
  niter = 200,
  threshold_data,
  omega_vec,
  p_mat,
  x_vec,
  meta_data,
  nbd_mat,
  gamma_fixed = FALSE,
  gamma = 0
)
}
\arguments{
\item{gamma_fixed}{}
}
\value{
assigned_domains

omega

gamma

pmat
}
\description{
Model for Subcellular Spatial Domain Discovery
}
