% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding_data.R
\name{threshold_data_function}
\alias{threshold_data_function}
\title{Threshold Binned Data}
\usage{
threshold_data_function(bin_data, threshold, meta_data)
}
\arguments{
\item{meta_data}{}
}
\value{
thresholded_data: A matrix of threshold data. Each gene is a binary vector where 1 represents those bins that have counts above the chosen threshold of that gene.

meta_data: Updated meta data
}
\description{
This function is used to threshold binned data.
This is an approach to reduce the impact of highly-expressed genes on the model outcomes.
}
\examples{
+
}
