% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts_data_with_assigned_domains.R
\name{transcripts_with_assigned_domains}
\alias{transcripts_with_assigned_domains}
\title{Assign domain to each transcript}
\usage{
transcripts_with_assigned_domains(meta_data, transcript_data, binsize)
}
\arguments{
\item{binsize}{}
}
\description{
This function merges the transcript data and the discovered domain data. All the transcripts in a bin are assigned the domain of that bin.
}
